Move firmware bundle validation tests to owner
